# openapi/openapi.v1.0.0-rc1.json
{
  "components": {
    "schemas": {
...
      "BacktestListResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BacktestRunResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "count"
        ],
        "title": "BacktestListResponse",
        "type": "object"
      },
      "BacktestRunRequest": {
        "properties": {
          "auto_analysis": {
            "default": true,
            "title": "Auto Analysis",
            "type": "boolean"
          },
          "bars": {
            "items": {
              "$ref": "#/components/schemas/BacktestBarRequest"
            },
            "title": "Bars",
            "type": "array"
          },
          "consult_execution_safety": {
            "default": true,
            "title": "Consult Execution Safety",
            "type": "boolean"
          },
          "fee_per_lot": {
            "default": "7",
            "title": "Fee Per Lot",
            "type": "string"
          },
          "initial_balance": {
            "default": "10000",
            "title": "Initial Balance",
            "type": "string"
          },
          "lot_size": {
            "default": "0.10",
            "title": "Lot Size",
            "type": "string"
          },
          "max_open_trades": {
            "default": 1,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Max Open Trades",
            "type": "integer"
          },
          "replay_mode": {
            "default": "candle",
            "description": "candle | tick",
            "title": "Replay Mode",
            "type": "string"
          },
          "request_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "slippage": {
            "default": "0.00005",
            "title": "Slippage",
            "type": "string"
          },
          "spread": {
            "default": "0.00010",
            "title": "Spread",
            "type": "string"
          },
          "stop_loss_distance": {
            "default": "0.0020",
            "title": "Stop Loss Distance",
            "type": "string"
          },
          "symbol": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Symbol",
            "type": "string"
          },
          "take_profit_distance": {
            "default": "0.0040",
            "title": "Take Profit Distance",
            "type": "string"
          },
          "ticks": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Ticks",
            "type": "array"
          },
          "timeframe": {
            "default": "m15",
            "maxLength": 16,
            "title": "Timeframe",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "symbol"
        ],
        "title": "BacktestRunRequest",
        "type": "object"
      },
      "BacktestRunResponse": {
        "properties": {
          "assumptions": {
            "additionalProperties": true,
            "title": "Assumptions",
            "type": "object"
          },
          "bar_count": {
            "default": 0,
            "title": "Bar Count",
            "type": "integer"
          },
          "equity_curve": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Equity Curve",
            "type": "array"
          },
          "error_message": {
            "default": "",
            "title": "Error Message",
            "type": "string"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "initial_balance": {
            "title": "Initial Balance",
            "type": "string"
          },
          "metrics": {
            "additionalProperties": true,
            "title": "Metrics",
            "type": "object"
          },
          "portfolio_snapshot": {
            "additionalProperties": true,
            "title": "Portfolio Snapshot",
            "type": "object"
          },
          "replay_mode": {
            "title": "Replay Mode",
            "type": "string"
          },
          "replay_state": {
            "additionalProperties": true,
            "title": "Replay State",
            "type": "object"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "timeframe": {
            "title": "Timeframe",
            "type": "string"
          },
          "trade_count": {
            "default": 0,
            "title": "Trade Count",
            "type": "integer"
          },
          "trades": {
            "items": {
              "$ref": "#/components/schemas/SimulatedTradeResponse"
            },
            "title": "Trades",
            "type": "array"
          }
        },
        "required": [
          "id",
          "request_id",
          "symbol",
          "timeframe",
          "status",
          "replay_mode",
          "initial_balance"
        ],
        "title": "BacktestRunResponse",
        "type": "object"
      },
...
      "BrokerDiagnosticsResponse": {
        "properties": {
          "broker_id": {
            "format": "uuid",
            "title": "Broker Id",
            "type": "string"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "connections": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Connections",
            "type": "array"
          },
          "discovered_capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Discovered Capabilities",
            "type": "array"
          },
          "last_error": {
            "default": "",
            "title": "Last Error",
            "type": "string"
          },
          "last_heartbeat_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Heartbeat At"
          },
          "last_successful_connection_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Successful Connection At"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "platform_code": {
            "title": "Platform Code",
            "type": "string"
          },
          "reconnect": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Reconnect",
            "type": "array"
          },
          "reconnect_count": {
            "default": 0,
            "title": "Reconnect Count",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "uptime_seconds": {
            "default": 0.0,
            "title": "Uptime Seconds",
            "type": "number"
          }
        },
        "required": [
          "broker_id",
          "status",
          "platform_code"
        ],
        "title": "BrokerDiagnosticsResponse",
        "type": "object"
      },
...
      "BrokerResponse": {
        "properties": {
          "broker_type": {
            "title": "Broker Type",
            "type": "string"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "country_code": {
            "title": "Country Code",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "platform_code": {
            "title": "Platform Code",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "website": {
            "title": "Website",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "broker_type",
          "status",
          "platform_code",
          "country_code",
          "website",
          "description",
          "created_at",
          "updated_at"
        ],
        "title": "BrokerResponse",
        "type": "object"
      },
...
      "CreateBrokerRequest": {
        "properties": {
          "activate": {
            "default": false,
            "title": "Activate",
            "type": "boolean"
          },
          "broker_type": {
            "default": "retail",
            "title": "Broker Type",
            "type": "string"
          },
          "capability_codes": {
            "items": {
              "type": "string"
            },
            "title": "Capability Codes",
            "type": "array"
          },
          "country_code": {
            "default": "US",
            "maxLength": 2,
            "minLength": 2,
            "title": "Country Code",
            "type": "string"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "platform_code": {
            "default": "other",
            "title": "Platform Code",
            "type": "string"
          },
          "slug": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Slug",
            "type": "string"
          },
          "website": {
            "default": "",
            "title": "Website",
            "type": "string"
          }
        },
        "required": [
          "name",
          "slug"
        ],
        "title": "CreateBrokerRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UpdateBrokerRequest": {
        "properties": {
          "broker_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Broker Type"
          },
          "country_code": {
            "anyOf": [
              {
                "maxLength": 2,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "platform_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform Code"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website"
          }
        },
        "title": "UpdateBrokerRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "QuantForg v1.0.0 Release Candidate 1 OpenAPI specification. EXECUTION_ENABLED remains false. No AI endpoints. Generated from FastAPI route registrations.",
    "title": "QuantForg API",
    "version": "1.0.0-rc.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/backtests": {
      "get": {
        "description": "List backtest runs for the current user.",
        "operationId": "list_backtests_api_v1_backtests_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Backtests",
        "tags": [
          "backtesting"
        ]
      }
    },
    "/api/v1/backtests/run": {
      "post": {
        "description": "Run an offline backtest. Never places live trades.",
        "operationId": "run_backtest_api_v1_backtests_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Run Backtest",
        "tags": [
          "backtesting"
        ]
      }
    },
    "/api/v1/backtests/{backtest_id}": {
      "get": {
        "description": "Get one backtest run with simulated trades and metrics.",
        "operationId": "get_backtest_api_v1_backtests__backtest_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "backtest_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Backtest Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Backtest",
        "tags": [
          "backtesting"
        ]
      }
    },
...
    "/api/v1/brokers": {
      "get": {
        "operationId": "list_brokers_api_v1_brokers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BrokerResponse"
                  },
                  "title": "Response List Brokers Api V1 Brokers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Brokers",
        "tags": [
          "brokers"
        ]
      },
      "post": {
        "operationId": "create_broker_api_v1_brokers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBrokerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrokerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Broker",
        "tags": [
          "brokers"
        ]
      }
    },
    "/api/v1/brokers/{broker_id}": {
      "delete": {
        "operationId": "delete_broker_api_v1_brokers__broker_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "broker_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Broker Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Broker",
        "tags": [
          "brokers"
        ]
      },
      "get": {
        "operationId": "get_broker_api_v1_brokers__broker_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "broker_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Broker Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrokerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Broker",
        "tags": [
          "brokers"
        ]
      },
      "patch": {
        "operationId": "update_broker_api_v1_brokers__broker_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "broker_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Broker Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBrokerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrokerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Broker",
        "tags": [
          "brokers"
        ]
      }
    },
    "/api/v1/brokers/{broker_id}/diagnostics": {
      "get": {
        "operationId": "get_broker_diagnostics_api_v1_brokers__broker_id__diagnostics_get",
        "parameters": [
          {
            "in": "path",
            "name": "broker_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Broker Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrokerDiagnosticsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Broker Diagnostics",
        "tags": [
          "brokers"
        ]
      }
    },
...
  }
}